refactor: Kotlin DSL 소개 기능 구현을 위한 객체들을 main 패키지로 분리
