+++ src/main/kotlin/study/Language.kt
@@ -1,0 +1,15 @@
+package study
+
+data class Language(val languages: Map<String, Int>)
+
+class LanguageBuilder {
+    private val languages: MutableMap<String, Int> = mutableMapOf()
+
+    infix fun String.level(value: Int) {
+        languages[this] = value
+    }
+
+    fun build(): Language {
+        return Language(languages)
+    }
+}
